backend/requirements.txt
beanie==1.24.0
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
argon2-cffi==21.3.0
python-multipart==0.0.6
